test: extract shared fakes into tests/fakes.py
