+++ tests/test_agstack_adapter.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from dataclasses import dataclass
 from datetime import date, datetime, time
 from uuid import UUID
 
@@ -23,30 +22,7 @@ from utils.agstack_adapter import (
     threatmodel_to_definition,
 )
 
-
-@dataclass
-class FakeCrop:
-    name: str
-
-
-@dataclass
-class FakeThreatModel:
-    scientific_name: str
-    common_name: str
-    crop: FakeCrop
-    definition: dict
-    threat_type: str | None = None
-
-
-@dataclass
-class FakeDataRow:
-    date: date
-    time: time
-    atmospheric_temperature: float | None = None
-    atmospheric_relative_humidity: float | None = None
-    precipitation: float | None = None
-    average_wind_speed: float | None = None
-    atmospheric_pressure: float | None = None
+from tests.fakes import FakeCrop, FakeDataRow, FakeThreatModel
 
 
 def _fungal_tm(**overrides) -> FakeThreatModel:
